Проведен небольшой рефакторинг
