+++ backend/src/routers/usersRouter.js
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+import { usersController } from '../controllers/usersController.js';
+
+export const usersRouter = Router();
+
+usersRouter.get('/', usersController.getAllUsers);
+
+usersRouter.get('/:userID', usersController.getUserByID);
+
+usersRouter.post('/', usersController.addNewUser);
